[Custom controls] Added color picker for iOS < 14
Might aswell add an option to optionally use this instead of the system
color picker on iOS 14+. Used library: Alderis

Other changes
- [Custom controls] The editor is now fullscreen due to the
presentation of the color picker.
- Fixes scroll wheel input is always int

+++ JavaApp/src/main/java/org/lwjgl/glfw/GLFW.java
@@ -1104,7 +1104,7 @@ public class GLFW
                         break;
                     case CallbackBridge.EVENT_TYPE_SCROLL:
                         if (mGLFWScrollCallback != null) {
-                            mGLFWScrollCallback.invoke(ptr, (int)dataArr[1], (int)dataArr[2]);
+                            mGLFWScrollCallback.invoke(ptr, (double)dataArr[1], (double)dataArr[2]);
                         }
                         break;
                     case CallbackBridge.EVENT_TYPE_WINDOW_POS:
